+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+1. Add `getPlatform` command to System plugin for distinguishing the current platform.
+
 ## [0.18.0] - 2026-08-24
 
 ### Changed